Build an artifact compatible with the protobuf gradle plugin
The protobuf gradle plugin supports downloading protoc plugins from
maven central:
https://github.com/google/protobuf-gradle-plugin#locate-external-executables.
This makes it much easier to use pb-and-k in a gradle build since the
`protoc-gen-kotlin` binary doesn't have to be compiled/installed
manually.

The new artifact is packaged as a self-executable jar using the Spring
Boot Gradle Plugin. This is required because `protoc` tries to execute
the provided artifact directly.