Add properties on the parent message for all fields inside of a oneof
This mimics the way oneof fields can be accessed in the protobuf Java
and Scala APIs.

For a proto message like:

```proto
message Foo {
    oneof type {
        string a_string = 1;
        Bar a_bar = 2;
    }
}
```

To access the oneof fields previously required a rather verbose
approach:

```kt
val a_string: String? = (foo.type as? Foo.Type.AString)?.value
```

This can now be simplified to:

```kt
val a_string: String? = foo.a_string
```

Which will return the value of the `a_string` field if the oneof is that
to that field, or otherwise return `null` if the oneof is set to a
different field or not set at all.