Fix two enum naming bugs with the new sealed class implementation
1. Since enum values now become `object` names in the generated code,
they're subject to the restrictions for Kotlin type names. This means
they can't be named after built-in types such as `String`. This was
causing a compiler error in `pbandk/wkt/descriptor.kt` because of the
`FieldDescriptorProto.Type.TYPE_STRING` enum value.

2. Enum values are now `object`s that inherit from a sealed class named
after the enum. The `object`'s name can't be the same as the sealed
class name or the compiler will complain of a circular reference. This
was causing a compiler error in `pbandk/wkt/struct.kt` because of the
`NullValue.NULL_VALUE` enum value.