Add detection of public API changes
The current runtime API will be compared against the contents of
`runtime/api/runtime.api` during `./gradlew build`. If there are any
changes, the build will fail. In order to explicitly confirm the
addition or removal of a public API, the `./gradlew apiDump` command
should be run. That will update the `runtime.api` file so it can
reviewed as part of the PR.

To make this work, I had to give each gradle project a distinct name in
`settings.gradle`. Otherwise the `ignoredProjects` property for the
`ApiValidationExtension` couldn't be configured properly.