Re-enable conformance tests that use JSON output
We accidentally forgot to re-enable these in
`pbandk.conformance.doTestIo()` after we added full JSON support to
pbandk.

Also enable the "recommended" tests since we're already passing most of
the ones that aren't related to JSON.

After enabling the tests, fixed a few bugs in the JSON code that the
tests exposed. The remaining failing tests were added to the
`failing_tests.txt` exception list and also to the checklist in
https://github.com/streem/pbandk/issues/72.