Publish pbandk on Bintray/JCenter
- Builds are configured to be published automatically by CI whenever a
git tag is created. They can also be published manually from the command
line as long as the `BINTRAY_API_USER` and `BINTRAY_API_KEY` environment
variables are defined.

- The maven group name has been updated to `pro.streem.pbandk` now that
we're not using Jitpack and don't need it to be tied to GitHub anymore.

- Because the Bintray gradle plugin hasn't been updated in a while and
does not work correctly with Kotlin Multiplatform projects, this PR just
uses the `maven-publish` gradle plugin directly.

- Also optimize build times in CI by disabling the gradle daemon and
enabling the build cache.