Fix JSON parsing of unknown enum values
Parsers using an older version of the enum definition would fail when
parsing data generated with a newer version of the enum definition. This
makes forwards-compatibility more brittle when dealing with enums.
According to discussion on
https://github.com/protocolbuffers/protobuf/issues/7392, the "ignore
unknown fields" parsing option should cause unknown enum values to be
ignored as well. This commit implements that behavior, based on the
implementation in protobuf-java at
https://github.com/protocolbuffers/protobuf/pull/4825.

The unit tests in `JsonTest` were also split out into separate files to
make them more manageable. The new tests are only those in `MapTest` and
`IgnoreUnknownFieldsTest`.