Fix performance problems when compiling generated code with Kotlin 1.4
When a protobuf message definition contains a lot of fields/oneofs, the
generated Kotlin code cannot be compiled by a Kotlin 1.4 compiler. The
compiler runs out of memory when compiling the long list of
`FieldDescriptor`s that gets passed to the `MessageDescriptor`. This
change implements a workaround that avoids the Kotlin compiler bug by
initializing the list of `FieldDescriptor`s outside of the
`MessageDescriptor` constructor.

Also add a test proto file to catch regressions with large protobuf
messages.

Fixes #94