Use GitHub Actions instead of CircleCI for PR tests
This migrates the CI jobs that get run pre-merge and post-merge for PRs.
The CI job for releasing a new version of pbandk to Bintray will be
migrated later.

With the update to Kotlin 1.4, we were hitting memory limits under
CircleCI. GitHub Actions has higher memory limits (and Streem also has a
paid account there).