Update to Kotlin 1.4
Aside from updates and simplifcations related to Kotlin 1.4, the other
major source of changes in this commit is the update of the
kotlinx-serialization library (required by the Kotlin 1.4 update). As
part of releasing v1.0, kotlinx-serialization made a bunch of changes to
their public API.

Also updated to newer versions of gradle, protobufjs, and the Spring
Boot Gradle Plugin.