Move Java test sources to a separate module
Rationale:both the Java and the Android plugin cannot be applied at the same time as they are incompatible:

- Jetbrains issue: https://youtrack.jetbrains.com/issue/KT-30878
- Android issue: https://issuetracker.google.com/issues/132576775

And because jvmTest contains Java sources, I had to move all this Java code to a separate module, where the Java plugin can be applied. This module is then. consumed as a dependency for jvmTest.