Allow reading multiple messages from an InputStream
Since protobuf messages are not self-delimiting, by default
`decodeFromStream()` will try to read until the end of the stream and
try to decode all bytes it reads as part of the message.  Applications
will often prefix a message with its length when writing multiple
messages to a single output stream. When consuming such a stream, the
application can read the length first and then pass it to
`decodeFromStream()` to make sure only that many bytes are read from the
stream.