Download protoc from maven for all protoc tasks
We no longer need to pass `protoc.path` to gradle, nor download and
compile `protoc` manually, when building pbandk. The protobuf source
does still need to be downloaded for running conformance tests since the
conformance test runner is not available as a maven artifact.

This functionality was already being used by the `DescriptorProtocTask`
in `protoc-gen-kotlin`. That task was a copy of the `ProtocTask` with
some changes made. This PR updates `ProtocTask` so it can be used as the
parent for `DescriptorProtocTask`.