Move the WKT .proto files to a separate pbandk-protos artifact
This allows `pbandk-runtime` to be used as a dependency in Android
gradle projects that also use the Protobuf Gradle Plugin. This used to
work correctly when the JVM and Android variants of `pbandk-runtime` had
a dependency on `protobuf-java(lite)`. Now that pbandk doesn't depend on
those libraries, pbandk has to provide the `.proto` files for the
Protocol Buffer Well Known Types.

The previous approach of adding the resources directory to the
`commonMain` sourceset had two problems:

1. It only seems to affect the `pbandk-runtime-jvm` jar. It does not add
the resources to the `pbandk-runtime-android` or
`pbandk-runtime-android-debug` aar files.

2. Even if the resources are added to the aar files, the Protobuf Gradle
Plugin has trouble consuming the .proto files properly because the aar
files are part of a Kotlin Multiplatform project.

To address both of these problems, the .proto files are now part of a
separate `pbandk-protos` artifact which gets built as a plain jar file.
By declaring `pbandk-protos` as a dependency of `pbandk-runtime-jvm` and
`pbandk-runtime-android(-debug)`, the Protobuf Gradle Plugin finds the
.proto files correctly even when the plugin is used in an Android
project.