Custom options polish
- Add a `Message.getFieldValue(fieldDescriptor)` method to simplify
reading a field value via protobuf reflection (e.g. when iterating over
the fields of a `Message` in a generic way).

- Make the use of `AtomicReference` an implementation detail by wrapping
it in an `ExtensionFieldSet` class.

- Annotate all public methods/interfaces that relate to protobuf custom
options with an `@ExperimentalProtoReflection` annotation since the API
is still in active development.

- Add `@get:pbandk.PbandkInternal` annotation to
`ExtendableMessage.extensionFields` property in generated code to work
around a bug in `binary-compatibility-validator` (see commend in the
code).

- Add `@PublicForGeneratedCode` annotation to various public APIs that
it was missing on and mark various public APIs `internal` that don't
actually need to be public.