Fix gradle metadata when building iOS/macOS builds separately (#178)
`pbandk-runtime` targets that are built and published on a different
machine than the main `pbandk-runtime` artifact were missing the
`pbandk-` prefix in the `pbandk-runtime` gradle metadata file that gets
published to Maven. This happens for example when `pbandk-runtime` is
built on a Linux host but `pbandk-runtime-iosx64/iosarm64/macosx64` are
built on a macOS host (such as in CI).

I believe this was happening because gradle does not create the
iOS/macOS publications when running on a Linux host, so the modification
of the `artifactId` to add the `pbandk-` prefix was getting skipped for
these publications. But the gradle metadata file would still include
references to these publications even though they weren't being created.

The solution ended up being pretty straightforward: instead of modifying
the publication, we can just modify the gradle project name to include
the `pbandk-` prefix. The project name is used as the default value for
the publication's `artifactId`.

Fixes #175