Refactor handling of special JSON types (#185)
Centralize all handling of protobuf messages that use a custom JSON representation into a new `JsonMessageAdapter` interface. This will make it easier to add types to the future without having to do lots of plumbing for each new type.

Note that this only includes protobuf messages that have a custom JSON representation but use the normal generated message in Kotlin. Protobuf messages that also use a custom type in Kotlin (namely, the `*Value` protobuf types that are converted to nullable primitive types in Kotlin) are not currently handled via `JsonMessageAdapter`.