Update pbandk-protos library to be Java 8 compatible
Without an explicit definition of the Java language version, gradle
will default to building artifacts that depend on the same version of
Java that gradle was running under. To guarantee that we don't break
compatibility with existing pbandk users, we need to explicitly tell
gradle to set the language version older than the currently-running Java
version.

Also update the example project to use Java 8 when building to ensure
that pbandk works correctly in a Java 8 project.

Fixes #211.