fix(enum): Use full import path in enum generation
The code generated for enums only had the full typename and not the full package + typename.
This caused errors when building the generated code if there are multiple proto fields with the same name since it could not resolve the references to the enum.

In order to fix this I made sure the generated code for enums uses the full import path.