Fix gradle task name for publishing to Sonatype
The full task name ends with `ToSonatypeRepository` rather than just
`ToSonatype`. The previously-used task name is now ambiguous (after some
of the changes made in the PR to update to Kotlin 1.9) and causes gradle
to fail with this error:

```
Task 'publishIosArm64PublicationToSonatype' is ambiguous in root project 'pbandk' and its subprojects. Candidates are: 'publishIosArm64PublicationToSonatypeRepository', 'publishIosSimulatorArm64PublicationToSonatypeRepository'.
```