Reduce chunk size for message descriptors in generated code
The previous size of 200 worked fine when a message contained lots of
fields without any custom options. But custom options on a field
increase the size of the code that is generated for the field's field
descriptor.  Which means we can fit fewer field descriptor constructor
calls in a single method before the method's bytecode becomes too big
and generates a `MethodTooLargeException`.

For now we'll just reduce the chunking size from 200 to 100 to give us
some more breathing room. Sufficiently large and complex protobuf
message definitions can still lead to an exception though. If that
happens often then we can implement further improvements to the
chunking.

Fixes #252